+++ parser_library/src/lsp/feature_provider.h
@@ -34,7 +34,7 @@ struct feature_provider
         position pos,
         char trigger_char,
         completion_trigger_kind trigger_kind) const = 0;
-    virtual document_symbol_list_s document_symbol(const std::string& document_uri) const = 0;
+    virtual document_symbol_list_s document_symbol(const std::string& document_uri, long long limit) const = 0;
 
 protected:
     ~feature_provider() = default;