+++ parser_library/src/workspace_manager.cpp
@@ -89,10 +89,21 @@ void workspace_manager::register_diagnostics_consumer(diagnostics_consumer* cons
     impl_->register_diagnostics_consumer(consumer);
 }
 
-void workspace_manager::register_performance_metrics_consumer(performance_metrics_consumer* consumer)
+void workspace_manager::unregister_diagnostics_consumer(diagnostics_consumer* consumer)
 {
-    impl_->register_performance_metrics_consumer(consumer);
+    impl_->unregister_diagnostics_consumer(consumer);
 }
+
+void workspace_manager::register_parsing_metadata_consumer(parsing_metadata_consumer* consumer)
+{
+    impl_->register_parsing_metadata_consumer(consumer);
+}
+
+void workspace_manager::unregister_parsing_metadata_consumer(parsing_metadata_consumer* consumer)
+{
+    impl_->unregister_parsing_metadata_consumer(consumer);
+}
+
 void workspace_manager::set_message_consumer(message_consumer* consumer) { impl_->set_message_consumer(consumer); }
 
 position_uri workspace_manager::definition(const char* document_uri, const position pos)
@@ -121,9 +132,9 @@ const std::vector<token_info>& workspace_manager::semantic_tokens(const char* do
     return impl_->semantic_tokens(document_uri);
 }
 
-document_symbol_list workspace_manager::document_symbol(const char* document_uri)
+document_symbol_list workspace_manager::document_symbol(const char* document_uri, long long limit)
 {
-    return impl_->document_symbol(document_uri);
+    return impl_->document_symbol(document_uri, limit);
 }
 
 } // namespace hlasm_plugin::parser_library